feat(core, common): move primary comparer implementations from `VecNotNan` to `F32NotNan`
